abstract Memory viewer
