+++ Source/ui_qt/DebugSupport/ELFSectionView.cpp
@@ -40,26 +40,7 @@ CELFSectionView::CELFSectionView(QMdiSubWindow* parent, QLayout* groupBoxLayout)
 		m_layout->addLayout(horizontalLayout);
 	}
 
-	m_model = new CQtMemoryViewModel(this);
-	m_memView = new QTableView(this);
-	m_memView->setModel(m_model);
-
-	QFont font = m_memView->font();
-	QFont fixedFont = QFontDatabase::systemFont(QFontDatabase::FixedFont);
-	fixedFont.setPointSize(font.pointSize());
-	m_memView->setFont(fixedFont);
-
-	QFontMetrics fmetric(fixedFont);
-	m_cwidth = fmetric.maxWidth();
-
-	auto header = m_memView->horizontalHeader();
-	header->setMinimumSectionSize(m_cwidth);
-	header->hide();
-	ResizeColumns();
-
-	m_memView->setContextMenuPolicy(Qt::CustomContextMenu);
-	connect(m_memView, &QTableView::customContextMenuRequested, this, &CELFSectionView::ShowContextMenu);
-
+	m_memView = new CMemoryViewTable(this);
 	auto verticalLayout = new QHBoxLayout();
 	verticalLayout->addWidget(m_memView);
 
@@ -97,146 +78,17 @@ void CELFSectionView::Reset()
 void CELFSectionView::showEvent(QShowEvent* evt)
 {
 	QWidget::showEvent(evt);
-	AutoColumn();
+	m_memView->ShowEvent();
 }
 
 void CELFSectionView::ResizeEvent()
 {
-	if(!m_model || m_bytesPerLine)
-		return;
-
-	AutoColumn();
-}
-
-void CELFSectionView::ShowContextMenu(const QPoint& pos)
-{
-	auto rightClickMenu = new QMenu(this);
-
-	{
-		auto bytesPerLineMenu = rightClickMenu->addMenu(tr("&Bytes Per Line"));
-		auto bytesPerLineActionGroup = new QActionGroup(this);
-		bytesPerLineActionGroup->setExclusive(true);
-
-		auto autoAction = bytesPerLineMenu->addAction("auto");
-		autoAction->setChecked(m_bytesPerLine == 0);
-		autoAction->setCheckable(true);
-		bytesPerLineActionGroup->addAction(autoAction);
-		connect(autoAction, &QAction::triggered, std::bind(&CELFSectionView::SetBytesPerLine, this, 0));
-
-		auto byte16Action = bytesPerLineMenu->addAction("16 bytes");
-		byte16Action->setChecked(m_bytesPerLine == 16);
-		byte16Action->setCheckable(true);
-		bytesPerLineActionGroup->addAction(byte16Action);
-		connect(byte16Action, &QAction::triggered, std::bind(&CELFSectionView::SetBytesPerLine, this, 16));
-
-		auto byte32Action = bytesPerLineMenu->addAction("32 bytes");
-		byte32Action->setChecked(m_bytesPerLine == 32);
-		byte32Action->setCheckable(true);
-		bytesPerLineActionGroup->addAction(byte32Action);
-		connect(byte32Action, &QAction::triggered, std::bind(&CELFSectionView::SetBytesPerLine, this, 32));
-	}
-
-	{
-		auto unitMenu = rightClickMenu->addMenu(tr("&Display Unit"));
-		auto unitActionGroup = new QActionGroup(this);
-		unitActionGroup->setExclusive(true);
-
-		auto activeUnit = m_model->GetActiveUnit();
-		for(uint32 i = 0; i < CQtMemoryViewModel::g_units.size(); i++)
-		{
-			const auto& unit = CQtMemoryViewModel::g_units[i];
-			auto itemAction = unitMenu->addAction(unit.description);
-			itemAction->setChecked(i == activeUnit);
-			itemAction->setCheckable(true);
-			unitActionGroup->addAction(itemAction);
-			connect(itemAction, &QAction::triggered, std::bind(&CELFSectionView::SetActiveUnit, this, i));
-		}
-	}
-
-	rightClickMenu->popup(m_memView->viewport()->mapToGlobal(pos));
-}
-
-void CELFSectionView::ResizeColumns()
-{
-	auto header = m_memView->horizontalHeader();
-	header->setSectionResizeMode(QHeaderView::Fixed);
-
-	auto units = m_model->UnitsForCurrentLine();
-	auto valueCell = m_cwidth * (m_model->CharsPerUnit() + 2);
-	int asciiCell = m_cwidth * (m_model->GetBytesPerUnit());
-	for(auto i = 0; i < units; ++i)
-	{
-		header->resizeSection(i, valueCell);
-	}
-	header->resizeSection(units, (units * asciiCell) + (m_cwidth * 2));
-
-	// collapse unused column
-	for(auto i = units + 1; i < m_maxUnits; ++i)
-	{
-		header->resizeSection(i, 0);
-	}
-	m_maxUnits = units + 1;
-}
-
-void CELFSectionView::AutoColumn()
-{
-	QFont font = m_memView->font();
-	QFontMetrics metric(font);
-	int columnHeaderWidth = metric.horizontalAdvance(" 0x00000000");
-
-	int tableWidth = m_memView->size().width() - columnHeaderWidth - m_memView->style()->pixelMetric(QStyle::PM_ScrollBarExtent);
-	auto bytesPerUnit = m_model->GetBytesPerUnit();
-
-	int valueCell = m_cwidth * (m_model->CharsPerUnit() + 2);
-	int asciiCell = m_cwidth * (m_model->GetBytesPerUnit());
-
-	int i = 0x2;
-	while(true)
-	{
-		int valueCellsWidth = i * valueCell;
-		int asciiCellsWidth = (i * asciiCell) + (m_cwidth * 2);
-		int totalWidth = valueCellsWidth + asciiCellsWidth;
-		if(totalWidth > tableWidth)
-		{
-			--i;
-			break;
-		}
-		++i;
-	}
-	m_model->SetUnitsForCurrentLine(i);
-	ResizeColumns();
-	m_model->Redraw();
-}
-
-void CELFSectionView::SetActiveUnit(int index)
-{
-	m_model->SetActiveUnit(index);
-	AutoColumn();
-}
-
-void CELFSectionView::SetSelectionStart(uint32 address)
-{
-	auto column = address % m_model->BytesForCurrentLine();
-	auto row = (address - column) / m_model->UnitsForCurrentLine();
-
-	auto index = m_model->index(row, column);
-	m_memView->selectionModel()->setCurrentIndex(index, QItemSelectionModel::ClearAndSelect);
-	m_memView->scrollTo(index, QAbstractItemView::PositionAtCenter);
+	m_memView->ResizeEvent();
 }
 
 void CELFSectionView::SetBytesPerLine(int bytesForLine)
 {
-	m_bytesPerLine = bytesForLine;
-	if(bytesForLine)
-	{
-		m_model->SetUnitsForCurrentLine(bytesForLine);
-		ResizeColumns();
-		m_model->Redraw();
-	}
-	else
-	{
-		AutoColumn();
-	}
+	m_memView->SetBytesPerLine(bytesForLine);
 }
 
 void CELFSectionView::SetELF(CELF* pELF)
@@ -341,7 +193,7 @@ void CELFSectionView::FillInformation(int section)
 	{
 		FillDynamicSectionListView(section);
 		m_dynSecTableWidget->show();
-		m_model->SetData(nullptr, 0);
+		m_memView->SetData(nullptr, 0);
 		m_memView->hide();
 	}
 	else if(pH->nType != CELF::SHT_NOBITS)
@@ -350,13 +202,13 @@ void CELFSectionView::FillInformation(int section)
 		auto getByte = [data](uint32 offset) {
 			return data[offset];
 		};
-		m_model->SetData(getByte, pH->nSize);
+		m_memView->SetData(getByte, pH->nSize);
 		m_memView->show();
 		m_dynSecTableWidget->hide();
 	}
 	else
 	{
-		m_model->SetData(nullptr, 0);
+		m_memView->SetData(nullptr, 0);
 		m_memView->show();
 		m_dynSecTableWidget->hide();
 	}
